Redux

+++ src/pages/Edit.jsx
@@ -28,11 +28,11 @@ export default function Edit() {
         <StButton
           type="button"
           onClick={() => {
-            const real = confirm("정말로 삭제하시겠습니까?");
-            if (real) {
+            const deletion = confirm("정말로 삭제하시겠습니까?");
+            if (deletion) {
               alert("삭제되었습니다.");
+              dispatch(deleteList(targetItem));
             }
-            dispatch(deleteList(targetItem));
           }}
         >
           삭제
